update Cardelivery còn sai  và thiếu 1 cái add

+++ Repository/Repositories/CarDeliveryHistoryRepository.cs
@@ -15,14 +15,32 @@ namespace Repository.Repositories
             _context = context;
         }
 
-        public async Task<IEnumerable<CarDeliveryHistory>> GetAllAsync()
+        public async Task<IEnumerable<CarDeliveryHistory>> GetAllAsync(int pageIndex, int pageSize)
         {
-            return await _context.CarDeliveryHistories.ToListAsync();
+            return await _context.CarDeliveryHistories
+                .Include(x => x.Car)
+                .Include(x => x.Customer)
+                .Include(x => x.Staff)
+                .Include(x => x.Location)
+                .OrderByDescending(x => x.DeliveryDate)
+                .Skip((pageIndex - 1) * pageSize)
+                .Take(pageSize)
+                .ToListAsync();
+        }
+
+        public async Task<int> CountAsync()
+        {
+            return await _context.CarDeliveryHistories.CountAsync();
         }
 
         public async Task<CarDeliveryHistory?> GetByIdAsync(int id)
         {
-            return await _context.CarDeliveryHistories.FindAsync(id);
+            return await _context.CarDeliveryHistories
+                .Include(x => x.Car)
+                .Include(x => x.Customer)
+                .Include(x => x.Staff)
+                .Include(x => x.Location)
+                .FirstOrDefaultAsync(x => x.Id == id);
         }
 
         public async Task AddAsync(CarDeliveryHistory entity)
@@ -37,15 +55,14 @@ namespace Repository.Repositories
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(CarDeliveryHistory entity)
+        public async Task DeleteAsync(int id)
         {
-            _context.CarDeliveryHistories.Remove(entity);
-            await _context.SaveChangesAsync();
+            var entity = await _context.CarDeliveryHistories.FindAsync(id);
+            if (entity != null)
+            {
+                _context.CarDeliveryHistories.Remove(entity);
+                await _context.SaveChangesAsync();
+            }
         }
-
-     
-
-
-      
     }
-}
+    }
